# Avoid a global lock and reference counting in `lockmap`