--- a/server/util/lockmap/lockmap.go
+++ b/server/util/lockmap/lockmap.go
@@ -2,38 +2,12 @@ package lockmap
 
 import (
 	"sync"
-	"sync/atomic"
 	"time"
 )
 
-type refCount struct {
-	i *int64
-}
-
-func (c *refCount) Inc() int64 {
-	j := atomic.AddInt64(c.i, 1)
-	return j
-}
-func (c *refCount) Dec() int64 {
-	j := atomic.AddInt64(c.i, -1)
-	return j
-}
-func (c *refCount) Val() int64 {
-	j := atomic.LoadInt64(c.i)
-	return j
-}
-
 type refCountedMutex struct {
 	sync.RWMutex
-	count refCount
-}
-
-func newRefCountedMutex() *refCountedMutex {
-	var i int64
-	return &refCountedMutex{
-		RWMutex: sync.RWMutex{},
-		count:   refCount{&i},
-	}
+	collected bool
 }
 
 // Locker implements a per-key mutex. This can be useful when protecting access
@@ -65,7 +39,6 @@ type Locker interface {
 
 type perKeyMutex struct {
 	mutexes sync.Map
-	bigLock sync.RWMutex
 	closed  chan struct{}
 }
 
@@ -87,46 +60,72 @@ func (p *perKeyMutex) gc() {
 		case <-t.C:
 		}
 		p.mutexes.Range(func(key, value any) bool {
-			p.bigLock.Lock()
 			rcm := value.(*refCountedMutex)
-			if rcm.count.Val() == 0 {
-				p.mutexes.Delete(key)
+			if !rcm.TryLock() {
+				// The mutex is in use and thus shouldn't be collected.
+				return true
 			}
-			p.bigLock.Unlock()
+			// At this point we hold the exclusive lock, so there are two cases:
+			// * There were no other goroutines waiting to acquire a lock, which
+			//   means that the mutex should be collected.
+			// * There were other goroutines waiting to acquire a lock, but they
+			//   failed to acquire it in the short period of time between it
+			//   becoming available and us acquiring it. As GC only rarely runs,
+			//   this case is expected to be very uncommon. All waiting
+			//   goroutines will need to refetch from the map.
+			//
+			// Mark the mutex as unusable before removing it from the map so
+			// that there are never two active mutexes for the same key.
+			rcm.collected = true
+			rcm.Unlock()
+			p.mutexes.Delete(key)
 			return true
 		})
 	}
 }
 
 func (p *perKeyMutex) Lock(key string) func() {
-	p.bigLock.Lock()
-	value, _ := p.mutexes.LoadOrStore(key, newRefCountedMutex())
-	rcm := value.(*refCountedMutex)
-	rcm.count.Inc()
-	p.bigLock.Unlock()
-
-	rcm.Lock()
-	return func() {
+	var rcm *refCountedMutex
+	for {
+		val, _ := p.mutexes.LoadOrStore(key, &refCountedMutex{})
+		rcm = val.(*refCountedMutex)
+		rcm.Lock()
+		if !rcm.collected {
+			break
+		}
+		// We hit the window between rcm.Unlock() and p.mutexes.Delete(key) in
+		// the gc() method. Busy wait for the map slot to be cleared.
 		rcm.Unlock()
-		rcm.count.Dec()
 	}
+	return func() { rcm.Unlock() }
 }
 
 func (p *perKeyMutex) RLock(key string) func() {
-	p.bigLock.Lock()
-	value, _ := p.mutexes.LoadOrStore(key, newRefCountedMutex())
-	rcm := value.(*refCountedMutex)
-	rcm.count.Inc()
-	p.bigLock.Unlock()
-
-	rcm.RLock()
-	return func() {
+	var rcm *refCountedMutex
+	for {
+		val, _ := p.mutexes.LoadOrStore(key, &refCountedMutex{})
+		rcm = val.(*refCountedMutex)
+		rcm.RLock()
+		if !rcm.collected {
+			break
+		}
+		// See comment in Lock().
 		rcm.RUnlock()
-		rcm.count.Dec()
 	}
+	return func() { rcm.RUnlock() }
 }
 
 func (p *perKeyMutex) Close() error {
 	close(p.closed)
 	return nil
+}
+
+// For testing only.
+func (p *perKeyMutex) count() int {
+	count := 0
+	p.mutexes.Range(func(key, value any) bool {
+		count++
+		return true
+	})
+	return count
 }