Instead of using a global lock on the entire `sync.Map` to avoid races with the custom GC routine, mark them as unusuable under their own write lock and have the `(R)Lock` function spin when it encounters such a mutex. This should only occur very rarely (GCs are performed at most once every 100ms) and the period of spinning should be short (as long as it takes to delete a key in the map). 

Reference counting is replaced with a call to `TryLock` for a particular entry, which combines a check for approximate unusedness with an acquire of the lock needed to mark the mutex as unusuable in a single atomic operation. This may succeed even though other goroutines were waiting on the lock, in which case they all have to refetch the new mutex from the map. This should also be a rare case since it requires GC to be performed at the exact time where the lock becomes available.

Benchmark:
```
goos: darwin
goarch: arm64
cpu: Apple M3 Max

before:
BenchmarkLockQPS-16     	 1000000	      1145.0 ns/op	     304 B/op	       7 allocs/op
BenchmarkRLockQPS-16    	 5469183	       210.8 ns/op	      72 B/op	       4 allocs/op

after:
BenchmarkLockQPS-16     	 1423785	       841.8 ns/op	     302 B/op	       6 allocs/op
BenchmarkRLockQPS-16    	62436318	        20.4 ns/op	      64 B/op	       3 allocs/op

            │   old.txt    │               new.txt               │
            │    sec/op    │   sec/op     vs base                │
LockQPS-16    1074.0n ± 2%   848.0n ± 1%  -21.05% (p=0.000 n=10)
RLockQPS-16   214.00n ± 4%   18.04n ± 3%  -91.57% (p=0.000 n=10)
geomean        479.4n        123.7n       -74.21%

            │  old.txt   │              new.txt               │
            │    B/op    │    B/op     vs base                │
LockQPS-16    305.0 ± 0%   302.0 ± 0%   -0.98% (p=0.000 n=10)
RLockQPS-16   72.00 ± 0%   64.00 ± 0%  -11.11% (p=0.000 n=10)
geomean       148.2        139.0        -6.18%

            │  old.txt   │              new.txt               │
            │ allocs/op  │ allocs/op   vs base                │
LockQPS-16    7.000 ± 0%   6.000 ± 0%  -14.29% (p=0.000 n=10)
RLockQPS-16   4.000 ± 0%   3.000 ± 0%  -25.00% (p=0.000 n=10)
geomean       5.292        4.243       -19.82%
```