@vanja-p 2025-05-15
You can use https://pkg.go.dev/golang.org/x/perf/cmd/benchstat to compare benchmark results. You might need to run you benchmarks with `--test.count=7` or something like that to use it.